requirements.txt
# ==================================================
# Customer Care Call Processing System - Dependencies
# ==================================================

# Google Drive API
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# AWS SDK
boto3==1.34.19
botocore==1.34.19

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.0
moto==4.2.12
black==23.12.1
flake8==7.0.0
mypy==1.8.0
boto3-stubs[essential,transcribe,bedrock-runtime,stepfunctions]==1.34.19

# Local testing
python-dotenv==1.0.0
localstack==3.0.0

# Type hints
types-requests==2.31.0
